For #15278: Moved increment function to extension function and fixed indentation
