+++ app/src/main/java/org/mozilla/fenix/FenixApplication.kt
@@ -137,9 +137,7 @@ open class FenixApplication : LocaleAwareApplication(), Provider {
             // to invoke parts of itself that require complete megazord initialization
             // before that process completes, we wait here, if necessary.
             if (!megazordSetup.isCompleted) {
-                runBlockingIncrement {
-                    megazordSetup.await()
-                }
+                runBlockingIncrement { megazordSetup.await() }
             }
         }
