+++ app/src/main/java/org/mozilla/fenix/perf/RunBlockingCounter.kt
@@ -12,6 +12,7 @@ package org.mozilla.fenix.perf
 import androidx.annotation.VisibleForTesting
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.runBlocking
+import org.mozilla.fenix.ext.getAndIncrementNoOverflow
 import java.util.concurrent.atomic.AtomicInteger
 import kotlin.coroutines.CoroutineContext
 
@@ -20,33 +21,22 @@ import kotlin.coroutines.CoroutineContext
  */
 @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
 object RunBlockingCounter {
-    var count = AtomicInteger(0)
-
-    fun incrementCounter() {
-        var prev: Int
-        var next: Int
-        do {
-            prev = count.get()
-            next = prev + 1
-            if (next == Integer.MAX_VALUE) next = 0
-        } while (!count.compareAndSet(prev, next))
-    }
+    val count = AtomicInteger(0)
 }
 
 /**
  * Wrapper around `runBlocking`. RunBlocking seems to be a "fix-all" to return values to the thread
  * where the coroutine is called. The official doc explains runBlocking:  "Runs a new coroutine and
- * blocks the current thread interruptibly until its completion`. This can have negative
- * side-effects on the our main thread which could lead to significant jank. This wrapper aims to
- * count the number of runBlocking call to try to limit them as much as possible to encourage
- * alternatives solutions whenever this function might be needed.
+ * blocks the current thread interruptibly until its completion`. This can block our main thread
+ * which could lead to significant jank. This wrapper aims to count the number of runBlocking call
+ * to try to limit them as much as possible to encourage alternatives solutions whenever this function
+ * might be needed.
  */
 fun <T> runBlockingIncrement(
     context: CoroutineContext? = null,
     action: suspend CoroutineScope.() -> T
 ): T {
-    RunBlockingCounter.incrementCounter()
-
+    RunBlockingCounter.count.getAndIncrementNoOverflow(RunBlockingCounter.count)
     return if (context != null) {
         runBlocking(context) { action() }
     } else {
