+++ app/src/test/java/org/mozilla/fenix/perf/RunBlockingCounterTest.kt
@@ -2,13 +2,11 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package org.mozilla.fenix
+package org.mozilla.fenix.perf
 
 import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Test
-import org.mozilla.fenix.perf.RunBlockingCounter
-import org.mozilla.fenix.perf.runBlockingIncrement
 
 class RunBlockingCounterTest {
 
@@ -18,7 +16,7 @@ class RunBlockingCounterTest {
     }
 
     @Test
-    fun `GIVEN we call runBlockingCounter THEN the count should increase`() {
+    fun `GIVEN we call our custom runBlocking method with counter THEN the latter should increase`() {
         assertEquals(0, RunBlockingCounter.count.get())
         runBlockingIncrement {}
         assertEquals(1, RunBlockingCounter.count.get())
