+++ mozilla-detekt-rules/src/main/java/org/mozilla/fenix/detektrules/MozillaRunBlockingCheck.kt
@@ -11,8 +11,8 @@ import org.jetbrains.kotlin.resolve.calls.callUtil.getCall
 import org.jetbrains.kotlin.resolve.calls.callUtil.getCalleeExpressionIfAny
 import kotlin.math.exp
 
-private const val VIOLATION_MSG = "Please use `org.mozilla.fenix.perf.RunBlockingCounter` instead " +
-        "because it allows us to monitor the code for performance regressions."
+private const val VIOLATION_MSG = "Please use `org.mozilla.fenix.perf.RunBlockingCounter.runBlockingIncrement`" +
+        "instead because it allows us to monitor the code for performance regressions."
 
 /**
  * A check to prevent us from working around mechanisms we implemented in
